Separate accept and initiate into own modules
