+++ p2panda-net/src/sync/mod.rs
@@ -1,782 +1,12 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+mod accept;
 mod config;
 mod handler;
+mod initiate;
 pub(crate) mod manager;
-
-use std::sync::Arc;
-
-use anyhow::Result;
-use futures_util::{AsyncRead, AsyncWrite, SinkExt};
-use iroh_net::key::PublicKey;
-use p2panda_sync::{FromSync, SyncError, SyncProtocol, Topic};
-use tokio::sync::{mpsc, oneshot};
-use tokio::task::JoinHandle;
-use tokio_util::sync::PollSender;
-use tracing::{debug, error, warn};
-
-use crate::engine::ToEngineActor;
-use crate::TopicId;
+#[cfg(test)]
+mod tests;
 
 pub use config::{ResyncConfiguration, SyncConfiguration};
 pub use handler::{SyncConnection, SYNC_CONNECTION_ALPN};
-
-/// Initiate a sync protocol session over the provided bi-directional stream for the given peer and
-/// topic.
-pub async fn initiate_sync<T, S, R>(
-    mut send: &mut S,
-    mut recv: &mut R,
-    peer: PublicKey,
-    topic: T,
-    sync_protocol: Arc<dyn for<'a> SyncProtocol<'a, T> + 'static>,
-    engine_actor_tx: mpsc::Sender<ToEngineActor<T>>,
-) -> Result<(), SyncError>
-where
-    T: Topic + TopicId + 'static,
-    S: AsyncWrite + Send + Unpin,
-    R: AsyncRead + Send + Unpin,
-{
-    debug!(
-        "initiate sync session with peer {} over topic {:?}",
-        peer, topic
-    );
-
-    engine_actor_tx
-        .send(ToEngineActor::SyncStart {
-            topic: topic.clone(),
-            peer,
-        })
-        .await
-        .map_err(|err| {
-            SyncError::Critical(format!("engine_actor_tx failed sending sync start: {err}"))
-        })?;
-
-    // Set up a channel for receiving messages from the sync session.
-    let (tx, mut rx) = mpsc::channel::<FromSync<T>>(128);
-    let mut sink = PollSender::new(tx).sink_map_err(|e| SyncError::Critical(e.to_string()));
-
-    // Spawn a "glue" task which represents the layer between the sync session and the engine.
-    //
-    // It picks up any messages from the sync session makes sure that the "Two-Phase Sync Flow" is
-    // followed (I. "Handshake" Phase & II. "Data Sync" Phase) and the engine accordingly informed
-    // about it.
-    //
-    // If the task detects any invalid behaviour of the sync flow, it fails critically, indicating
-    // that the sync protocol implementation does not behave correctly and is not compatible with
-    // the engine.
-    //
-    // Additionally, the task forwards any synced application data straight to the engine.
-    let glue_task_handle: JoinHandle<Result<(), SyncError>> = {
-        let mut sync_handshake_success = false;
-        let topic = topic.clone();
-
-        tokio::spawn(async move {
-            while let Some(message) = rx.recv().await {
-                // I. Handshake Phase.
-                //
-                // At the beginning of every sync session the "initiating" peer needs to send over
-                // the topic to the "accepting" peer during the handshake phase. This is the first
-                // message we're expecting:
-                if let FromSync::HandshakeSuccess(_) = message {
-                    // Receiving the handshake message twice is a protocol violation.
-                    if sync_handshake_success {
-                        return Err(SyncError::Critical(
-                            "received handshake message twice from sync session in handshake phase"
-                                .into(),
-                        ));
-                    }
-                    sync_handshake_success = true;
-
-                    // Inform the engine that we are expecting sync messages from the peer on this
-                    // topic.
-                    engine_actor_tx
-                        .send(ToEngineActor::SyncHandshakeSuccess {
-                            peer,
-                            topic: topic.clone(),
-                        })
-                        .await
-                        .map_err(|err| {
-                            SyncError::Critical(format!(
-                                "engine_actor_tx failed sending sync handshake success: {err}"
-                            ))
-                        })?;
-
-                    continue;
-                }
-
-                // 2. Data Sync Phase.
-                // ~~~~~~~~~~~~~~~~~~~
-                let FromSync::Data(header, payload) = message else {
-                    return Err(SyncError::Critical("expected to receive only data messages from sync session in data sync phase".into()));
-                };
-
-                engine_actor_tx
-                    .send(ToEngineActor::SyncMessage {
-                        header,
-                        payload,
-                        delivered_from: peer,
-                        topic: topic.clone(),
-                    })
-                    .await
-                    .map_err(|err| {
-                        SyncError::Critical(format!(
-                            "engine_actor_tx failed sending sync message: {err}"
-                        ))
-                    })?;
-            }
-
-            engine_actor_tx
-                .send(ToEngineActor::SyncDone { peer, topic })
-                .await
-                .map_err(|err| {
-                    SyncError::Critical(format!("engine_actor_tx failed sending sync done: {err}"))
-                })?;
-
-            Ok(())
-        })
-    };
-
-    // Run the "initiating peer" side of the sync protocol.
-    let result = sync_protocol
-        .initiate(
-            topic,
-            Box::new(&mut send),
-            Box::new(&mut recv),
-            Box::new(&mut sink),
-        )
-        .await;
-
-    // Drop the tx, so the rx in the glue task receives the closing event.
-    drop(sink);
-
-    // We're expecting the task to exit with a result soon, we're awaiting it here ..
-    let glue_task_result = glue_task_handle
-        .await
-        .map_err(|err| SyncError::Critical(format!("glue task handle failed: {err}")))?;
-
-    // .. to inform some brave developer who will read the error logs ..
-    if let Err(SyncError::Critical(err)) = &glue_task_result {
-        error!("critical error in sync protocol: {err}");
-    }
-
-    // .. and forward it further.
-    glue_task_result?;
-
-    // The same we're doing with errors coming from the sync protocol implementation itself.
-    if let Err(err) = result {
-        match &err {
-            SyncError::Critical(err) => {
-                error!("critical error in sync protocol: {err}");
-            }
-            _ => {
-                warn!("error in sync protocol: {err}");
-            }
-        }
-        return Err(err);
-    }
-
-    Ok(())
-}
-
-/// Accept a sync protocol session over the provided bi-directional stream for the given peer and
-/// topic.
-pub async fn accept_sync<T, S, R>(
-    mut send: &mut S,
-    mut recv: &mut R,
-    peer: PublicKey,
-    sync_protocol: Arc<dyn for<'a> SyncProtocol<'a, T> + 'static>,
-    engine_actor_tx: mpsc::Sender<ToEngineActor<T>>,
-) -> Result<(), SyncError>
-where
-    T: Topic + TopicId + 'static,
-    S: AsyncWrite + Send + Unpin,
-    R: AsyncRead + Send + Unpin,
-{
-    debug!("accept sync session with peer {}", peer);
-
-    // Set up a channel for receiving messages from the sync session.
-    let (tx, mut rx) = mpsc::channel::<FromSync<T>>(128);
-    let mut sink = PollSender::new(tx).sink_map_err(|e| SyncError::Critical(e.to_string()));
-
-    // Set up a channel for sending over errors to the "glue" task which occurred during sync.
-    let (sync_error_tx, mut sync_error_rx) = oneshot::channel::<SyncError>();
-
-    // Spawn a "glue" task which represents the layer between the sync session and the engine.
-    //
-    // It picks up any messages from the sync session makes sure that the "Two-Phase Sync Flow" is
-    // followed (I. "Handshake" Phase & II. "Data Sync" Phase) and the engine accordingly informed
-    // about it.
-    //
-    // If the task detects any invalid behaviour of the sync flow, it fails critically, indicating
-    // that the sync protocol implementation does not behave correctly and is not compatible with
-    // the engine.
-    //
-    // Additionally, the task forwards any synced application data straight to the engine.
-    let glue_task_handle: JoinHandle<Result<(), SyncError>> = tokio::spawn(async move {
-        let mut topic = None;
-
-        loop {
-            tokio::select! {
-                biased;
-
-                Ok(err) = &mut sync_error_rx => {
-                    engine_actor_tx
-                        .send(ToEngineActor::SyncFailed {
-                            peer,
-                            topic: topic.clone(),
-                        })
-                        .await
-                        .map_err(|err| {
-                            SyncError::Critical(
-                                format!("engine_actor_tx failed sending sync failed: {err}")
-                            )
-                        })?;
-
-                    // If we're observing an error we terminate the task here and propagate that
-                    // error further up.
-                    return Err(err);
-                },
-                message = rx.recv() => {
-                    let Some(message) = message else {
-                        // Sink (tx) got dropped, so we're leaving the task.
-                        break;
-                    };
-
-                    // I. Handshake Phase.
-                    //
-                    // At the beginning of every sync session the "accepting" peer needs to learn
-                    // the topic of the "initiating" peer during the handshake phase. This is
-                    // _always_ the first message we're expecting:
-                    if let FromSync::HandshakeSuccess(handshake_topic) = message {
-                        // It should only be sent once so topic should be `None` now.
-                        if topic.is_some() {
-                            return Err(
-                                SyncError::Critical(
-                                    "received topic twice from sync session in handshake phase"
-                                    .into()
-                                )
-                            );
-                        }
-
-                        topic = Some(handshake_topic.clone());
-
-                        // Inform the engine that we are expecting sync messages from the peer on
-                        // this topic.
-                        engine_actor_tx
-                            .send(ToEngineActor::SyncHandshakeSuccess {
-                                peer,
-                                topic: handshake_topic,
-                            })
-                            .await
-                            .map_err(|err| {
-                                SyncError::Critical(
-                                    format!("engine_actor_tx failed sending handshake success: {err}")
-                                )
-                            })?;
-
-                        continue;
-                    }
-
-                    // II. Data Sync Phase.
-                    //
-                    // At this stage we're beginning the actual "sync" protocol and expect messages
-                    // containing the data which was received from the "initiating" peer.
-                    //
-                    // Please note that in not all sync implementations the "accepting" peers
-                    // receives data.
-
-                    // The topic must be known at this point in order to process further messages.
-                    //
-                    // Any sync protocol implementation should have already failed with an
-                    // "unexpected behaviour" error if the topic wasn't learned. If this didn't
-                    // happen (due to an incorrect implementation) we will critically fail now.
-                    let Some(topic) = &topic else {
-                        return Err(
-                            SyncError::Critical(
-                                "never received topic from sync session in handshake phase"
-                                .into()
-                            )
-                        );
-                    };
-
-                    // From this point on we are only expecting "data" messages from the sync
-                    // session.
-                    let FromSync::Data(header, payload) = message else {
-                        return Err(
-                            SyncError::Critical(
-                                "expected only data messages from sync session in data sync phase"
-                                .into()
-                            )
-                        );
-                    };
-
-                    engine_actor_tx
-                        .send(ToEngineActor::SyncMessage {
-                            header,
-                            payload,
-                            delivered_from: peer,
-                            topic: topic.clone(),
-                        })
-                        .await
-                        .map_err(|err| {
-                            SyncError::Critical(
-                                format!("engine_actor_tx failed sending sync message: {err}")
-                            )
-                        })?;
-                },
-            }
-        }
-
-        // If topic was never set then we didn't receive any messages. In that case, the engine
-        // wasn't ever informed, so we can return here silently.
-        let Some(topic) = topic else {
-            return Ok(());
-        };
-
-        engine_actor_tx
-            .send(ToEngineActor::SyncDone { peer, topic })
-            .await
-            .map_err(|err| {
-                SyncError::Critical(format!("engine_actor_tx failed sending sync done: {err}"))
-            })?;
-
-        Ok(())
-    });
-
-    // Run the "accepting peer" side of the sync protocol.
-    let result = sync_protocol
-        .accept(
-            Box::new(&mut send),
-            Box::new(&mut recv),
-            Box::new(&mut sink),
-        )
-        .await;
-
-    // Drop the tx, so the rx in the glue task receives the closing event.
-    drop(sink);
-
-    // The sync protocol failed and we're informing the "glue" task about it, so it can accordingly
-    // wind down and inform the engine about it.
-    if let Err(sync_session_err) = result {
-        sync_error_tx
-            .send(sync_session_err)
-            .map_err(|err| SyncError::Critical(format!("sync_error_tx failed: {err}")))?;
-    }
-
-    // We're expecting the task to exit with a result soon, we're awaiting it here ..
-    let glue_task_result = glue_task_handle
-        .await
-        .map_err(|err| SyncError::Critical(format!("glue task handle failed: {err}")))?;
-
-    // .. to inform some brave developer who will read the error logs ..
-    if let Err(SyncError::Critical(err)) = &glue_task_result {
-        error!("critical error in sync protocol: {err}");
-    }
-
-    // .. and forward it further!
-    glue_task_result
-}
-
-#[cfg(test)]
-mod sync_protocols {
-    use std::sync::Arc;
-
-    use async_trait::async_trait;
-    use futures_lite::{AsyncRead, AsyncWrite, StreamExt};
-    use futures_util::{Sink, SinkExt};
-    use p2panda_sync::cbor::{into_cbor_sink, into_cbor_stream};
-    use p2panda_sync::{FromSync, SyncError, SyncProtocol};
-    use serde::{Deserialize, Serialize};
-
-    use super::tests::TestTopic;
-
-    #[derive(Debug, Serialize, Deserialize)]
-    enum ProtocolMessage {
-        Topic(TestTopic),
-        Done,
-    }
-
-    /// A sync implementation which returns a mocked error.
-    #[derive(Debug)]
-    pub enum FailingProtocol {
-        /// A critical error is triggered inside `accept()` after sync messages have been
-        /// exchanged.
-        AcceptorFailsCritical,
-
-        /// A critical error is triggered inside `initiate()` after the handshake is complete.
-        InitiatorFailsCritical,
-
-        /// An unexpected behaviour error is triggered inside `accept()` by sending the topic twice
-        /// from `initiate()`.
-        InitiatorSendsTopicTwice,
-
-        /// An unexpected behaviour error is triggered inside `initiate()` by sending a topic from
-        /// `accept()`.
-        AcceptorSendsTopic,
-
-        /// No errors are explicitly triggered; used for "happy path" test.
-        NoError,
-    }
-
-    #[async_trait]
-    impl<'a> SyncProtocol<'a, TestTopic> for FailingProtocol {
-        fn name(&self) -> &'static str {
-            "failing-protocol"
-        }
-
-        async fn initiate(
-            self: Arc<Self>,
-            topic: TestTopic,
-            tx: Box<&'a mut (dyn AsyncWrite + Send + Unpin)>,
-            rx: Box<&'a mut (dyn AsyncRead + Send + Unpin)>,
-            mut app_tx: Box<
-                &'a mut (dyn Sink<FromSync<TestTopic>, Error = SyncError> + Send + Unpin),
-            >,
-        ) -> Result<(), SyncError> {
-            let mut sink = into_cbor_sink(tx);
-            let mut stream = into_cbor_stream(rx);
-
-            sink.send(ProtocolMessage::Topic(topic.clone())).await?;
-
-            // Simulate critical sync implementation bug by sending the topic a second time.
-            if let FailingProtocol::InitiatorSendsTopicTwice = *self {
-                sink.send(ProtocolMessage::Topic(topic.clone())).await?;
-            }
-
-            sink.send(ProtocolMessage::Done).await?;
-
-            app_tx.send(FromSync::HandshakeSuccess(topic)).await?;
-
-            // Simulate some critical error which occurred inside the sync session.
-            if let FailingProtocol::InitiatorFailsCritical = *self {
-                return Err(SyncError::Critical(
-                    "something really bad happened in the initiator".to_string(),
-                ));
-            }
-
-            while let Some(result) = stream.next().await {
-                let message: ProtocolMessage = result?;
-                match &message {
-                    ProtocolMessage::Topic(_) => {
-                        return Err(SyncError::UnexpectedBehaviour(
-                            "unexpected message received from acceptor".to_string(),
-                        ));
-                    }
-                    ProtocolMessage::Done => break,
-                }
-            }
-
-            Ok(())
-        }
-
-        async fn accept(
-            self: Arc<Self>,
-            tx: Box<&'a mut (dyn AsyncWrite + Send + Unpin)>,
-            rx: Box<&'a mut (dyn AsyncRead + Send + Unpin)>,
-            mut app_tx: Box<
-                &'a mut (dyn Sink<FromSync<TestTopic>, Error = SyncError> + Send + Unpin),
-            >,
-        ) -> Result<(), SyncError> {
-            let mut sink = into_cbor_sink(tx);
-            let mut stream = into_cbor_stream(rx);
-
-            // Simulate critical sync implementation bug by sending the topic from the acceptor (it
-            // _never_ sends any topics).
-            if let FailingProtocol::AcceptorSendsTopic = *self {
-                let topic = TestTopic::new("unexpected behaviour test");
-                sink.send(ProtocolMessage::Topic(topic)).await?;
-            }
-
-            let mut received_topic = false;
-
-            while let Some(result) = stream.next().await {
-                let message: ProtocolMessage = result?;
-                match &message {
-                    ProtocolMessage::Topic(topic) => {
-                        if !received_topic {
-                            app_tx
-                                .send(FromSync::HandshakeSuccess(topic.clone()))
-                                .await?;
-                            received_topic = true;
-                        } else {
-                            return Err(SyncError::UnexpectedBehaviour(
-                                "received topic too often".to_string(),
-                            ));
-                        }
-                    }
-                    ProtocolMessage::Done => break,
-                }
-            }
-
-            if let FailingProtocol::AcceptorFailsCritical = *self {
-                return Err(SyncError::Critical(
-                    "something really bad happened in the acceptor".to_string(),
-                ));
-            }
-
-            sink.send(ProtocolMessage::Done).await?;
-
-            Ok(())
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::sync::Arc;
-
-    use iroh_net::NodeId;
-    use p2panda_core::{Hash, PrivateKey};
-    use p2panda_sync::{SyncError, Topic};
-    use serde::{Deserialize, Serialize};
-    use tokio::sync::mpsc;
-    use tokio::task::JoinHandle;
-    use tokio_util::compat::{TokioAsyncReadCompatExt, TokioAsyncWriteCompatExt};
-
-    use crate::engine::ToEngineActor;
-    use crate::{sync, TopicId};
-
-    use super::sync_protocols::FailingProtocol;
-
-    #[derive(Clone, Debug, PartialEq, Eq, Hash, Deserialize, Serialize)]
-    pub struct TestTopic(String, [u8; 32]);
-
-    impl TestTopic {
-        pub fn new(name: &str) -> Self {
-            Self(name.to_owned(), *Hash::new(&name).as_bytes())
-        }
-    }
-
-    impl Topic for TestTopic {}
-
-    impl TopicId for TestTopic {
-        fn id(&self) -> [u8; 32] {
-            self.1
-        }
-    }
-
-    async fn run_sync_impl(
-        protocol: FailingProtocol,
-    ) -> (
-        mpsc::Receiver<ToEngineActor<TestTopic>>,
-        mpsc::Receiver<ToEngineActor<TestTopic>>,
-        JoinHandle<Result<(), SyncError>>,
-        JoinHandle<Result<(), SyncError>>,
-    ) {
-        let topic = TestTopic::new("run test protocol impl");
-
-        let initiator_node_id =
-            NodeId::from_bytes(PrivateKey::new().public_key().as_bytes()).unwrap();
-        let acceptor_node_id =
-            NodeId::from_bytes(PrivateKey::new().public_key().as_bytes()).unwrap();
-
-        let sync_protocol = Arc::new(protocol);
-
-        // Duplex streams which simulate both ends of a bi-directional network connection.
-        let (initiator_stream, acceptor_stream) = tokio::io::duplex(64 * 1024);
-        let (initiator_read, initiator_write) = tokio::io::split(initiator_stream);
-        let (acceptor_read, acceptor_write) = tokio::io::split(acceptor_stream);
-
-        // Channel for sending messages out of a running sync session.
-        let (initiator_tx, initiator_rx) = mpsc::channel(128);
-        let (acceptor_tx, acceptor_rx) = mpsc::channel(128);
-
-        let sync_protocol_clone = sync_protocol.clone();
-
-        let initiator_handle = {
-            let topic = topic.clone();
-
-            tokio::spawn(async move {
-                sync::initiate_sync(
-                    &mut initiator_write.compat_write(),
-                    &mut initiator_read.compat(),
-                    acceptor_node_id,
-                    topic.clone(),
-                    sync_protocol,
-                    initiator_tx,
-                )
-                .await
-            })
-        };
-
-        let acceptor_handle = {
-            tokio::spawn(async move {
-                sync::accept_sync(
-                    &mut acceptor_write.compat_write(),
-                    &mut acceptor_read.compat(),
-                    initiator_node_id,
-                    sync_protocol_clone,
-                    acceptor_tx,
-                )
-                .await
-            })
-        };
-
-        (initiator_rx, acceptor_rx, initiator_handle, acceptor_handle)
-    }
-
-    #[tokio::test]
-    async fn initiator_fails_critical() {
-        let (mut rx_initiator, mut rx_acceptor, initiator_handle, acceptor_handle) =
-            run_sync_impl(FailingProtocol::InitiatorFailsCritical).await;
-
-        // Expected initiator messages.
-        assert!(matches!(
-            rx_initiator.recv().await,
-            Some(ToEngineActor::SyncStart { .. })
-        ));
-
-        assert!(matches!(
-            rx_initiator.recv().await,
-            Some(ToEngineActor::SyncHandshakeSuccess { .. })
-        ));
-
-        assert!(matches!(
-            rx_initiator.recv().await,
-            Some(ToEngineActor::SyncDone { .. })
-        ));
-
-        // @TODO: Where is the SyncFail message?
-
-        // Expected acceptor messages.
-        assert!(matches!(
-            rx_acceptor.recv().await,
-            Some(ToEngineActor::SyncHandshakeSuccess { .. })
-        ));
-
-        assert!(matches!(
-            rx_acceptor.recv().await,
-            Some(ToEngineActor::SyncDone { .. })
-        ));
-
-        // Expected handler results.
-        assert_eq!(
-            initiator_handle.await.unwrap(),
-            Err(SyncError::Critical(
-                "something really bad happened in the initiator".into(),
-            ))
-        );
-        assert_eq!(acceptor_handle.await.unwrap(), Ok(()));
-    }
-
-    #[tokio::test]
-    async fn initiator_sends_topic_twice() {
-        let (mut _rx_initiator, mut _rx_acceptor, initiator_handle, acceptor_handle) =
-            run_sync_impl(FailingProtocol::InitiatorSendsTopicTwice).await;
-
-        assert_eq!(initiator_handle.await.unwrap(), Ok(()));
-        assert_eq!(
-            acceptor_handle.await.unwrap(),
-            // This is _not_ a critical error as the acceptor protocol implementation handled the
-            // protocol violation (sending topic twice) by itself.
-            Err(SyncError::UnexpectedBehaviour(
-                "received topic too often".into(),
-            ))
-        );
-    }
-
-    #[tokio::test]
-    async fn acceptor_fails_critical() {
-        let (mut rx_initiator, mut _rx_acceptor, initiator_handle, acceptor_handle) =
-            run_sync_impl(FailingProtocol::AcceptorFailsCritical).await;
-
-        // Expected initiator messages.
-        assert!(matches!(
-            rx_initiator.recv().await,
-            Some(ToEngineActor::SyncStart { .. })
-        ));
-
-        assert!(matches!(
-            rx_initiator.recv().await,
-            Some(ToEngineActor::SyncHandshakeSuccess { .. })
-        ));
-
-        assert!(matches!(
-            rx_initiator.recv().await,
-            Some(ToEngineActor::SyncDone { .. })
-        ));
-
-        // Expected handler results.
-        assert_eq!(initiator_handle.await.unwrap(), Ok(()));
-        assert_eq!(
-            acceptor_handle.await.unwrap(),
-            Err(SyncError::Critical(
-                "something really bad happened in the acceptor".into(),
-            ))
-        );
-    }
-
-    #[tokio::test]
-    async fn acceptor_sends_topic() {
-        let (mut rx_initiator, mut _rx_acceptor, initiator_handle, acceptor_handle) =
-            run_sync_impl(FailingProtocol::AcceptorSendsTopic).await;
-
-        // Expected initiator messages.
-        assert!(matches!(
-            rx_initiator.recv().await,
-            Some(ToEngineActor::SyncStart { .. })
-        ));
-
-        assert!(matches!(
-            rx_initiator.recv().await,
-            Some(ToEngineActor::SyncHandshakeSuccess { .. })
-        ));
-
-        assert!(matches!(
-            rx_initiator.recv().await,
-            Some(ToEngineActor::SyncDone { .. })
-        ));
-
-        // Expected handler results.
-        assert_eq!(
-            initiator_handle.await.unwrap(),
-            Err(SyncError::UnexpectedBehaviour(
-                "unexpected message received from acceptor".into(),
-            ))
-        );
-        assert_eq!(acceptor_handle.await.unwrap(), Ok(()));
-    }
-
-    #[tokio::test]
-    async fn run_sync_without_error() {
-        let (mut rx_initiator, mut rx_acceptor, initiator_handle, acceptor_handle) =
-            run_sync_impl(FailingProtocol::NoError).await;
-
-        // Expected initiator messages.
-        assert!(matches!(
-            rx_initiator.recv().await,
-            Some(ToEngineActor::SyncStart { .. })
-        ));
-
-        assert!(matches!(
-            rx_initiator.recv().await,
-            Some(ToEngineActor::SyncHandshakeSuccess { .. })
-        ));
-
-        assert!(matches!(
-            rx_initiator.recv().await,
-            Some(ToEngineActor::SyncDone { .. })
-        ));
-
-        // Expected acceptor messages.
-        // @TODO: Is there a `SyncStart` message for the acceptor?
-        // assert!(matches!(
-        //     rx_acceptor.recv().await,
-        //     Some(ToEngineActor::SyncStart { .. })
-        // ));
-
-        assert!(matches!(
-            rx_acceptor.recv().await,
-            Some(ToEngineActor::SyncHandshakeSuccess { .. })
-        ));
-
-        assert!(matches!(
-            rx_acceptor.recv().await,
-            Some(ToEngineActor::SyncDone { .. })
-        ));
-
-        // Expected handler results.
-        assert_eq!(initiator_handle.await.unwrap(), Ok(()));
-        assert_eq!(acceptor_handle.await.unwrap(), Ok(()));
-    }
-}
